Change Read() size and return value to a size_t for clarity
BUG=100641



TEST=N/A
Review URL: http://codereview.chromium.org/8352019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106533 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/filters/ffmpeg_demuxer.cc
@@ -368,7 +368,7 @@ base::TimeDelta FFmpegDemuxer::GetStartTime() const {
   return start_time_;
 }
 
-int FFmpegDemuxer::Read(int size, uint8* data) {
+size_t FFmpegDemuxer::Read(size_t size, uint8* data) {
   DCHECK(data_source_);
 
   // If read has ever failed, return with an error.
